+++ desktop/cmp/input/JsonInput.js
@@ -5,23 +5,16 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 
-import {codeInput} from '@xh/hoist/desktop/cmp/input/CodeInput';
-import {elemFactory, HoistComponent, LayoutSupport} from '@xh/hoist/core';
-import {bindable} from '@xh/hoist/mobx';
-import PT from 'prop-types';
-
 import {HoistInput} from '@xh/hoist/cmp/input';
-
+import {elemFactory, HoistComponent, LayoutSupport} from '@xh/hoist/core';
+import {codeInput, CodeInput} from '@xh/hoist/desktop/cmp/input';
+import {bindable} from '@xh/hoist/mobx';
 import * as codemirror from 'codemirror';
-import {jsonlint} from './impl/jsonlint';
-
 import 'codemirror/mode/javascript/javascript';
+import {jsonlint} from './impl/jsonlint';
 
 /**
  * Code-editor style input for editing and validating JSON, powered by CodeMirror.
- *
- * TODO - understanding sizing spec / requirements for component vs. generated CodeMirror.
- * Reconcile LayoutSupport with width/height props. https://github.com/xh/hoist-react/issues/327
  */
 @HoistComponent
 @LayoutSupport
@@ -30,20 +23,7 @@ export class JsonInput extends HoistInput {
     @bindable fullScreen = false;
 
     static propTypes = {
-        ...HoistInput.propTypes,
-        value: PT.string,
-
-        /** True to commit on every change/keystroke, default false. */
-        commitOnChange: PT.bool,
-
-        /**
-         * Configuration object with any properties supported by the CodeMirror API.
-         * @see {@link https://codemirror.net/doc/manual.html#api_configuration|CodeMirror Docs}
-         */
-        editorProps: PT.object,
-
-        /** True to show Fullscreen + Auto-format buttons at top-right of input. */
-        showActionButtons: PT.bool
+        ...CodeInput.propTypes
     };
 
     baseClassName = 'xh-json-input';
@@ -54,31 +34,28 @@ export class JsonInput extends HoistInput {
             linter: this.linter,
             formatter: this.formatter,
             mode: 'application/json',
-            showFullscreenButton: this.showActionButtons,
-            showFormatButton: this.showActionButtons,
             ...props
         });
     }
 
     linter(text) {
-        const found = [];
+        const errors = [];
+        if (!text) return errors;
 
         jsonlint.parseError = function(str, hash) {
             const loc = hash.loc;
-            found.push({
+            errors.push({
                 from: codemirror.Pos(loc.first_line - 1, loc.first_column),
                 to: codemirror.Pos(loc.last_line - 1, loc.last_column),
                 message: str
             });
         };
 
-        if (!text) return found;
-
         try {
             jsonlint.parse(text);
         } catch (ignored) {}
 
-        return found;
+        return errors;
     }
 
     formatter(text) {